+++ src/main/java/module-info.java
@@ -3,8 +3,9 @@ module org.schlunzis.jduino {
     requires org.slf4j;
 
     exports org.schlunzis.jduino;
-    exports org.schlunzis.jduino.proto.tlv;
+    exports org.schlunzis.jduino.protocol.tlv;
     exports org.schlunzis.jduino.simple;
-    exports org.schlunzis.jduino.proto;
+    exports org.schlunzis.jduino.protocol;
     exports org.schlunzis.jduino.channel.serial;
+    exports org.schlunzis.jduino.channel;
 }
