--- src/main/java/org/schlunzis/jduino/ChannelFactory.java
@@ -1,10 +1,0 @@
-package org.schlunzis.jduino;
-
-import org.schlunzis.jduino.proto.Protocol;
-
-@FunctionalInterface
-public interface ChannelFactory<P extends Protocol<P>, C extends Channel<P>> {
-
-    C createChannel(P protocol);
-
-}
