--- src/main/java/org/schlunzis/jduino/CommandType.java
@@ -1,7 +1,0 @@
-package org.schlunzis.jduino;
-
-public interface CommandType {
-
-    byte getCode();
-
-}
