+++ src/main/java/org/schlunzis/jduino/channel/ChannelMessageListener.java
@@ -1,7 +1,7 @@
-package org.schlunzis.jduino;
+package org.schlunzis.jduino.channel;
 
-import org.schlunzis.jduino.proto.Message;
-import org.schlunzis.jduino.proto.Protocol;
+import org.schlunzis.jduino.protocol.Message;
+import org.schlunzis.jduino.protocol.Protocol;
 
 @FunctionalInterface
 public interface ChannelMessageListener<P extends Protocol<P>> {
