--- src/main/java/org/schlunzis/jduino/proto/plain/PlainDataListener.java
@@ -1,4 +1,0 @@
-package org.schlunzis.jduino.proto.plain;
-
-public class PlainDataListener {
-}
