+++ src/main/java/org/schlunzis/jduino/protocol/MessageDecoder.java
@@ -1,4 +1,4 @@
-package org.schlunzis.jduino.proto;
+package org.schlunzis.jduino.protocol;
 
 public interface MessageDecoder<P extends Protocol<P>> {
