+++ src/main/java/org/schlunzis/jduino/protocol/MessageEncoder.java
@@ -1,4 +1,4 @@
-package org.schlunzis.jduino.proto;
+package org.schlunzis.jduino.protocol;
 
 public interface MessageEncoder<P extends Protocol<P>> {
