+++ src/main/java/org/schlunzis/jduino/protocol/plain/PlainDataListener.java
@@ -1,0 +1,4 @@
+package org.schlunzis.jduino.protocol.plain;
+
+public class PlainDataListener {
+}
